Changes to RequestURLBuilder to only include relevant fields in API response. Moves RequestOptions to entities.
